requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0

# RAG / Vector Search
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM APIs
groq==0.4.2
anthropic==0.18.1

# Data Processing
numpy==1.26.4
pandas==2.2.2

# Visualization
plotly==5.18.0

# Utilities
requests==2.31.0
tqdm==4.66.1